external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/extensions/webgl-debug-renderer-info-expected.txt

This test verifies the functionality of the WEBGL_debug_renderer_info extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing enums with extension disabled
PASS getError was expected value: INVALID_ENUM : UNMASKED_VENDOR_WEBGL should not be queryable if extension is disabled
PASS getError was expected value: INVALID_ENUM : UNMASKED_RENDERER_WEBGL should not be queryable if extension is disabled
PASS No WEBGL_debug_renderer_info support -- this is legal
PASS WEBGL_debug_renderer_info not listed as supported and getExtension failed -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
